Start fixing tests for combobox and select
