+++ packages/components/src/Form/Inputs/Combobox/Combobox.test.tsx
@@ -30,202 +30,18 @@ import React from 'react'
 
 import { Combobox, ComboboxInput, ComboboxList, ComboboxOption } from '.'
 
-// for the requestAnimationFrame in handleBlur (not working currently)
-// jest.useFakeTimers()
-
 afterEach(cleanup)
-
-describe('<Combobox/> with options', () => {
-  test('with handleChange', () => {
-    const options = [{ value: 'FOO' }, { value: 'BAR' }]
-    const handleChange = jest.fn()
-    const {
-      // getAllByRole,
-      queryByText,
-      getByText,
-      getByPlaceholderText,
-    } = renderWithTheme(
-      <Combobox
-        options={options}
-        id="with-options"
-        inputProps={{ placeholder: 'Search' }}
-        onChange={handleChange}
-      />
-    )
-    expect(queryByText('FOO')).not.toBeInTheDocument()
-    expect(queryByText('BAR')).not.toBeInTheDocument()
-
-    const input = getByPlaceholderText('Search')
-    expect(input).toBeVisible()
-
-    act(() => {
-      fireEvent.click(input)
-    })
-
-    // const foo = getByText('FOO')
-    const bar = getByText('BAR')
-
-    // Clicking on the options should fire onBlur on the input and
-    // trigger the state transition that allows in an updated input value.
-    // It doesn't, and the following doesn't work to fake it, so we can't test input value.
-    // act(() => {
-    // getAllByRole('option')[0].focus()
-    // input.blur()
-    // })
-    fireEvent.click(bar)
-    // fireEvent.click(foo)
-
-    expect(handleChange).toHaveBeenCalledTimes(1)
-    // expect(handleChange).toHaveBeenCalledWith({ value: 'FOO' })
-    expect(handleChange).toHaveBeenCalledWith({ value: 'BAR' })
-  })
-
-  test('with openOnFocus', () => {
-    const options = [{ value: 'FOO' }, { value: 'BAR' }]
-    const { getByRole, queryByRole, getByPlaceholderText } = renderWithTheme(
-      <Combobox
-        options={options}
-        id="with-options"
-        inputProps={{ placeholder: 'Search' }}
-        openOnFocus
-      />
-    )
-
-    expect(queryByRole('listbox')).not.toBeInTheDocument()
-
-    getByPlaceholderText('Search').focus()
-    expect(getByRole('listbox')).toBeInTheDocument()
-  })
-
-  describe('Keyboard navigation', () => {
-    const arrowDown = {
-      key: 'ArrowDown',
-    }
-    const arrowUp = {
-      key: 'ArrowUp',
-    }
-    const enter = {
-      key: 'Enter',
-    }
-
-    test('arrows and enter', () => {
-      const options = [{ value: 'FOO' }, { value: 'BAR' }]
-      const {
-        getAllByRole,
-        getByRole,
-        queryByRole,
-        getByPlaceholderText,
-      } = renderWithTheme(
-        <Combobox
-          options={options}
-          id="with-options"
-          inputProps={{ placeholder: 'Search' }}
-          openOnFocus
-        />
-      )
-
-      expect(queryByRole('listbox')).not.toBeInTheDocument()
-
-      const input = getByPlaceholderText('Search')
-
-      fireEvent.keyDown(input, arrowDown)
-      expect(getByRole('listbox')).toBeInTheDocument()
-
-      const items = getAllByRole('option')
-      expect(input).toHaveValue('')
-      expect(items[0]).not.toHaveAttribute('aria-selected')
-      expect(items[1]).not.toHaveAttribute('aria-selected')
-
-      fireEvent.keyDown(input, arrowDown)
-      expect(input).toHaveValue('FOO')
-      expect(items[0]).toHaveAttribute('aria-selected', 'true')
-      expect(items[1]).toHaveAttribute('aria-selected', 'false')
-
-      fireEvent.keyDown(input, arrowDown)
-      expect(input).toHaveValue('BAR')
-      expect(items[0]).toHaveAttribute('aria-selected', 'false')
-      expect(items[1]).toHaveAttribute('aria-selected', 'true')
-
-      fireEvent.keyDown(input, arrowDown)
-      expect(input).toHaveValue('')
-      expect(items[0]).not.toHaveAttribute('aria-selected')
-      expect(items[1]).not.toHaveAttribute('aria-selected')
-
-      fireEvent.keyDown(input, arrowUp)
-      expect(input).toHaveValue('BAR')
-      expect(items[0]).toHaveAttribute('aria-selected', 'false')
-      expect(items[1]).toHaveAttribute('aria-selected', 'true')
-
-      fireEvent.keyDown(input, enter)
-      expect(input).toHaveValue('BAR')
-      expect(queryByRole('listbox')).not.toBeInTheDocument()
-    })
-
-    test('arrows and enter with autoComplete = false', () => {
-      const options = [{ value: 'FOO' }, { value: 'BAR' }]
-      const {
-        getAllByRole,
-        getByRole,
-        queryByRole,
-        getByPlaceholderText,
-      } = renderWithTheme(
-        <Combobox
-          options={options}
-          id="with-options"
-          inputProps={{ autocomplete: false, placeholder: 'Search' }}
-          openOnFocus
-        />
-      )
-
-      expect(queryByRole('listbox')).not.toBeInTheDocument()
-
-      const input = getByPlaceholderText('Search')
-
-      fireEvent.keyDown(input, arrowDown)
-      expect(getByRole('listbox')).toBeInTheDocument()
-
-      const items = getAllByRole('option')
-      expect(input).toHaveValue('')
-      expect(items[0]).not.toHaveAttribute('aria-selected')
-      expect(items[1]).not.toHaveAttribute('aria-selected')
-
-      fireEvent.keyDown(input, arrowDown)
-      expect(input).toHaveValue('')
-      expect(items[0]).toHaveAttribute('aria-selected', 'true')
-      expect(items[1]).toHaveAttribute('aria-selected', 'false')
-
-      fireEvent.keyDown(input, arrowDown)
-      expect(input).toHaveValue('')
-      expect(items[0]).toHaveAttribute('aria-selected', 'false')
-      expect(items[1]).toHaveAttribute('aria-selected', 'true')
-
-      fireEvent.keyDown(input, arrowDown)
-      expect(input).toHaveValue('')
-      expect(items[0]).toHaveAttribute('aria-selected', 'true')
-      expect(items[1]).toHaveAttribute('aria-selected', 'false')
-
-      fireEvent.keyDown(input, arrowUp)
-      expect(input).toHaveValue('')
-      expect(items[0]).toHaveAttribute('aria-selected', 'false')
-      expect(items[1]).toHaveAttribute('aria-selected', 'true')
-
-      fireEvent.keyDown(input, enter)
-      expect(input).toHaveValue('BAR')
-      expect(queryByRole('listbox')).not.toBeInTheDocument()
-    })
-  })
-})
 
 describe('<Combobox/> with children', () => {
   test('Renders children, merges callbacks', () => {
     const handleChange = jest.fn()
     const handleClick = jest.fn()
     const { getByText, getByTestId, queryByText } = renderWithTheme(
-      <Combobox options={[{ value: 'Apples' }]} onChange={handleChange}>
+      <Combobox onChange={handleChange}>
         <ComboboxInput data-testid="select-input" />
         <ComboboxList>
-          <ComboboxOption value="Foo" data={101} onClick={handleClick} />
-          <ComboboxOption value="Bar" data={102} />
+          <ComboboxOption label="Foo" value="101" onClick={handleClick} />
+          <ComboboxOption label="Bar" value="102" />
         </ComboboxList>
       </Combobox>
     )
